Encode a label to a value with other metadata

The most fundamental components of a `code` object are the `label` and `value` elements. A `code` object is essentially a key-value tuple that has some extra metadata.

code(
  label,
  value,
  description = label,
  links_from = label,
  missing = FALSE,
  ...
)

Arguments

label
A short label for a value in a vector

value
The raw value found in the respective vector

description
A longer-form label for the value, if extra context for that label is needed

links_from
A reference to another `code` in another `coding` object for recoding purposes.

missing
Whether this `code` represents a missing response

...
Any extra metadata